requirements.txt
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorlog==6.7.0
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
googledrivedownloader==0.4
idna==3.4
imageio==2.27.0
importlib-resources==5.12.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.22.3
psutil==5.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rdflib==6.3.2
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.19.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==16.7.6
wandb==0.14.2
yacs==0.1.8
zipp==3.15.0
